Rename and refactor util
