+++ util/bitmap_test.go
